@@ -6,7 +6,7 @@ import 'package:webview_flutter/webview_flutter.dart';
 class ArticleDetailPage extends StatefulWidget {
   final String articleUrl;
 
-  ArticleDetailPage({@required this.articleUrl});
+  ArticleDetailPage({required this.articleUrl});
 
   @override
   _ArticleDetailPageState createState() => _ArticleDetailPageState();
@@ -25,12 +25,13 @@ class _ArticleDetailPageState extends State<ArticleDetailPage> {
       body: Container(
         height: MediaQuery.of(context).size.height,
         width: MediaQuery.of(context).size.width,
-        child: WebView(
+        child: Text(widget.articleUrl),
+        /*child: WebView(
           initialUrl: widget.articleUrl,
           onWebViewCreated: (WebViewController webViewController) {
             _controller.complete(webViewController);
           },
-        ),
+        ),*/
       ),
     );
   }
